fixup! Posts page (#3)
