+++ src/components/search-bar/SearchBar.jsx
@@ -6,7 +6,7 @@ const SearchBar = ({ name, label, value, onChange, className, placeholder }) => 
     <div className={className}>
       <label htmlFor={name}>{label}</label>
       <input
-        className='bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-indigo-500 focus:border-indigo-500 block w-full  p-2.5'
+        className='bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-indigo-500 focus:border-indigo-500 block w-full p-2.5'
         type='search'
         name={name}
         id={name}
