+++ src/components/spinner/SpinnerFullHeight.jsx
@@ -2,7 +2,7 @@ import Spinner from '@app/components/spinner/Spinner';
 import withLogger, { getLoggerProps } from '@app/utils/logger/withLogger';
 
 const SpinnerFullHeight = props => (
-  <div className='min-h-[length:var(--main-layout-height)] min-w-full mt-12 flex items-center justify-center'>
+  <div className='min-h-[length:var(--main-layout-height)] min-w-full flex items-center justify-center'>
     <Spinner {...getLoggerProps(props)} size={48} />
   </div>
 );
